Fix typings II
